+++ index.js
@@ -1,10 +1,11 @@
 const express = require("express");
 const app = express();
 const port = 5000;
-const routes = require('./routes/twitter')
+const routes = require('./routes/twitter');
+const twitterLogin = require('./routes/TwitterLogin');
+const userData = require("./routes/UserData");
 const cookieParser = require("cookie-parser");
 const session = require("express-session");
-
 
 main().catch((err) => console.error(err.message, err));
 
@@ -14,6 +15,8 @@ async function main() {
 
   app.use(require("body-parser").urlencoded({ extended: true }));
   app.use('', routes);
+  app.use('', twitterLogin);
+  app.use('', userData);
 
   app.set("views", __dirname + "/views");
   app.set("view engine", "ejs");
